requirements.txt
sanic==22.6.2
numpy==1.21.6
lmdb==1.2.1
opencv-python==4.6.0.66
PyYAML==5.4.1
tqdm==4.62.2
yapf==0.31.0
basicsr==1.4.2
facexlib==0.2.5
gfpgan==1.3.8
Pillow
requests_futures
realesrgan
